Add QueryReader to read data/queries.txt
queries.txt contains the WHERE-clauses of all queries of interest.
Where the queries basically consist of a conjunctive predicate whose
clauses are made up of: a selected attribute compared to a value from
the attribute's domain (see Tuple.java), using the less-than or
greater-than operator.
Therefore, the QueryReader reads each query of the file and creates
a Query-object which contains a set of SelectionOperators, each
consisting of a key, an operator and a value.